# server_side/application/static/swagger.yaml
openapi: 3.0.0
info:
  title: Masters_Swagger
  contact: {}
  version: "1.0"
servers:
  - url: http:/127.0.0.1:5000
    variables: {}
paths:
...
  /post/:
    post:
      tags:
        - Post
      summary: Add
      operationId: Add
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddRequest"
            example:
              content_image:
              description: description_test
              style_id: 0
        required: true
      responses:
        200:
          description: "OK"
          headers: {}
      deprecated: false
  /post/{id}:
    get:
      tags:
        - Post
      summary: Get
      operationId: Get
      parameters:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: integer
            format: int32
            example: 2
      responses:
        200:
          description: "OK"
          headers: {}
      deprecated: false

    put:
      tags:
        - Post
      summary: Update
      operationId: Update
      parameters:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: integer
            format: int32
            example: 6
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRequest"
            example:
              content_image:
              description: description_test_update
        required: true
      responses:
        200:
          description: "OK"
          headers: {}
      deprecated: false

    delete:
      tags:
        - Post
      summary: Delete
      operationId: Delete
      parameters:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: integer
            format: int32
            example: 14
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteRequest"
            example:
              login: filip
              password: haslo
              email: fifi
        required: true
      responses:
        200:
          description: "OK"
          headers: {}
      deprecated: false
  /user/{id}/posts:
    get:
      tags:
        - Post
      summary: GetAll
      operationId: GetAll
      parameters:
        - name: limit
          in: query
          description: ""
          required: true
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 10
        - name: page
          in: query
          description: ""
          required: true
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 1
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: integer
            format: int32
            example: 7
      responses:
        200:
          description: "OK"
          headers: {}
      deprecated: false
  /user/followed/posts:
    get:
      tags:
        - Post
      summary: GetFollowedPosts
      operationId: GetFollowedPosts
      parameters:
        - name: limit
          in: query
          description: ""
          required: true
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 10
        - name: page
          in: query
          description: ""
          required: true
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 1
      responses:
        200:
          description: "OK"
          headers: {}
      deprecated: false
  /post/{id}/like:
    put:
      tags:
        - Post
      summary: LikePost
      operationId: LikePost
      parameters:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: integer
            format: int32
            example: 8
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LikePostRequest"
            example:
              content_image:
              description: description_test_update
        required: true
      responses:
        200:
          description: "OK"
          headers: {}
      deprecated: false
  /post/{id}/unlike:
    put:
      tags:
        - Post
      summary: UnLikePost
      operationId: UnLikePost
      parameters:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: integer
            format: int32
            example: 8
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnLikePostRequest"
            example:
              content_image:
              description: description_test_update
        required: true
      responses:
        200:
          description: "OK"
          headers: {}
      deprecated: false
  /post/{id}/likes:
    get:
      tags:
        - Post
      summary: GetLikesList
      operationId: GetLikesList
      parameters:
        - name: limit
          in: query
          description: ""
          required: true
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 10
        - name: page
          in: query
          description: ""
          required: true
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 1
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: integer
            format: int32
            example: 8
      responses:
        200:
          description: "OK"
          headers: {}
      deprecated: false
...
components:
  schemas:
...
    AddRequest:
      title: AddRequest
      required:
        - content_image
        - description
        - style_id
      type: object
      properties:
        content_image:
          type: string
          nullable: true
        description:
          type: string
        style_id:
          type: integer
          format: int32
      example:
        content_image:
        description: description_test
        style_id: 0
    UpdateRequest:
      title: UpdateRequest
      required:
        - content_image
        - description
      type: object
      properties:
        content_image:
          type: string
          nullable: true
        description:
          type: string
      example:
        content_image:
        description: description_test_update
    DeleteRequest:
      title: DeleteRequest
      required:
        - login
        - password
        - email
      type: object
      properties:
        login:
          type: string
        password:
          type: string
        email:
          type: string
      example:
        login: filip
        password: haslo
        email: fifi
    LikePostRequest:
      title: LikePostRequest
      required:
        - content_image
        - description
      type: object
      properties:
        content_image:
          type: string
          nullable: true
        description:
          type: string
      example:
        content_image:
        description: description_test_update
    UnLikePostRequest:
      title: UnLikePostRequest
      required:
        - content_image
        - description
      type: object
      properties:
        content_image:
          type: string
          nullable: true
        description:
          type: string
      example:
        content_image:
        description: description_test_update
...
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
  - httpBearer: []
...